Initial general validations consolidated
