+++ src/js/edu-benefits/utils/validations.js
@@ -2,38 +2,21 @@ import _ from 'lodash';
 import moment from 'moment';
 import { states } from './options-for-select';
 import { showRelinquishedEffectiveDate } from './helpers';
-import { isValidDateOver17, isBlankAddress, isValidPartialDateField, isValidYear, isValidPartialMonthYearInPast } from '../../common/utils/validations';
 import { dateToMoment } from '../../common/utils/helpers';
-
-function validateIfDirty(field, validator) {
-  if (field.dirty) {
-    return validator(field.value);
-  }
-
-  return true;
-}
-
-function validateIfDirtyDate(dayField, monthField, yearField, validator) {
-  if (dayField.dirty && monthField.dirty && yearField.dirty) {
-    return validator(dayField.value, monthField.value, yearField.value);
-  }
-
-  return true;
-}
-
-function validateIfDirtyDateObj(date, validator) {
-  return validateIfDirtyDate(date.day, date.month, date.year, () => {
-    return validator(date);
-  });
-}
-
-function isBlank(value) {
-  return value === '';
-}
-
-function isNotBlank(value) {
-  return value !== '';
-}
+import {
+  isBlank,
+  isBlankAddress,
+  isNotBlank,
+  isValidDateField,
+  isValidDateOver17,
+  isValidEmail,
+  isValidName,
+  isValidPartialDateField,
+  isValidPartialMonthYearInPast,
+  isValidPhone,
+  isValidSSN,
+  isValidYear
+} from '../../common/utils/validations';
 
 function isValidYearOrBlank(value) {
   return isValidYear(value) || value === '';
@@ -47,79 +30,11 @@ function isValidMonths(value) {
   return Number(value) >= 0;
 }
 
-// Conditions for valid SSN from the original 1010ez pdf form:
-// '123456789' is not a valid SSN
-// A value where the first 3 digits are 0 is not a valid SSN
-// A value where the 4th and 5th digits are 0 is not a valid SSN
-// A value where the last 4 digits are 0 is not a valid SSN
-// A value with 3 digits, an optional -, 2 digits, an optional -, and 4 digits is a valid SSN
-// 9 of the same digits (e.g., '111111111') is not a valid SSN
-function isValidSSN(value) {
-  if (value === '123456789' || value === '123-45-6789') {
-    return false;
-  } else if (/^0{3}-?\d{2}-?\d{4}$/.test(value)) {
-    return false;
-  } else if (/^\d{3}-?0{2}-?\d{4}$/.test(value)) {
-    return false;
-  } else if (/^\d{3}-?\d{2}-?0{4}$/.test(value)) {
-    return false;
-  }
-
-  const noBadSameDigitNumber = _.without(_.range(0, 10), 2, 4, 5)
-    .every(i => {
-      const sameDigitRegex = new RegExp(`${i}{3}-?${i}{2}-?${i}{4}`);
-      return !sameDigitRegex.test(value);
-    });
-
-  if (!noBadSameDigitNumber) {
-    return false;
-  }
-
-  return /^\d{3}-?\d{2}-?\d{4}$/.test(value);
-}
-
-function isValidDate(day, month, year) {
-  // Use the date class to see if the date parses back sanely as a
-  // validation check. Not sure is a great idea...
-  const adjustedMonth = Number(month) - 1;  // JS Date object 0-indexes months. WTF.
-  const date = new Date(year, adjustedMonth, day);
-  const today = new Date();
-
-  if (today < date) {
-    return false;
-  }
-
-  if (Number(year) < 1900) {
-    return false;
-  }
-
-  return date.getDate() === Number(day) &&
-    date.getMonth() === adjustedMonth &&
-    date.getFullYear() === Number(year);
-}
-
-function isValidName(value) {
-  return /^[a-zA-Z][a-zA-Z '\-]*$/.test(value);
-}
-
 function isValidMonetaryValue(value) {
   if (value !== null) {
     return /^[$]{0,1}\d+\.?\d*$/.test(value);
   }
   return true;
-}
-
-// TODO: look into validation libraries (npm "validator")
-function isValidPhone(value) {
-  // Strip spaces, dashes, and parens
-  const stripped = value.replace(/[^\d]/g, '');
-  // Count number of digits
-  return /^\d{10}$/.test(stripped);
-}
-
-function isValidEmail(value) {
-  // Comes from StackOverflow: http://stackoverflow.com/questions/46155/validate-email-address-in-javascript
-  return /^(([^<>()\[\]\\.,;:\s@"]+(\.[^<>()\[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/.test(value);
 }
 
 // Pulled from https://en.wikipedia.org/wiki/Routing_transit_number#Check_digit
@@ -158,10 +73,6 @@ function isBlankMonthYear(field) {
 
 function isNotBlankDateField(field) {
   return isNotBlank(field.day.value) && isNotBlank(field.month.value) && isNotBlank(field.year.value);
-}
-
-function isValidDateField(field) {
-  return isValidDate(field.day.value, field.month.value, field.year.value);
 }
 
 function isValidFutureDate(day, month, year) {
@@ -404,25 +315,13 @@ function isValidPage(completePath, pageData) {
 }
 
 export {
-  validateIfDirty,
-  validateIfDirtyDate,
-  validateIfDirtyDateObj,
-  isBlank,
-  isNotBlank,
   isNotBlankDateField,
-  isValidDate,
-  isValidName,
-  isValidSSN,
   isValidMonetaryValue,
-  isValidPhone,
-  isValidEmail,
-  isValidYear,
   isValidYearOrBlank,
   isValidCurrentOrPastYear,
   isValidMonths,
   isValidRoutingNumber,
   isValidField,
-  isValidDateField,
   isValidFutureOrPastDateField,
   isValidDateRange,
   isValidForm,
